Block Controller bring up. Moved cllback code to node, block control is next. Fixed a bug in CDC PWM.

+++ LcsBaseStation/main.cpp
@@ -252,7 +252,7 @@ uint8_t bsRefreshActiveSessionCallback( ) {
 // station itself issues such a request, the defined callback is invoked.
 //
 //------------------------------------------------------------------------------------------------------------
-uint8_t lcsReqCallback( uint8_t npId, uint8_t item, uint16_t *arg1, uint16_t *arg2 ) {
+uint8_t lcsReqCallback( uint16_t npId, uint8_t item, uint16_t *arg1, uint16_t *arg2 ) {
 
     printf( "REQ callback: npId: 0x%x, item: %d", npId, item );
     if ( arg1 != nullptr ) printf( ", arg1: %d, ", *arg1 ); else printf( ", arg1: null" );
@@ -264,7 +264,7 @@ uint8_t lcsReqCallback( uint8_t npId, uint8_t item, uint16_t *arg1, uint16_t *ar
 // When the base station gets a reply message for a request previously sent, this callback is invoked.
 //
 //------------------------------------------------------------------------------------------------------------
-uint8_t lcsRepCallback( uint8_t npId, uint8_t item, uint16_t arg1, uint16_t arg2, uint8_t ret ) {
+uint8_t lcsRepCallback( uint16_t npId, uint8_t item, uint16_t arg1, uint16_t arg2, uint8_t ret ) {
 
     printf( "REP callback: npId: 0x%x, item: %d, arg1: %d, arg2: %d, ret: %d ", npId, item , arg1, arg2, ret );
     return( ALL_OK );
